src/indexer/requirements.txt
azure-servicebus==7.12.2
fastapi==0.112.0
python-dotenv==1.0.1
azure-storage-blob==12.22.0
python-multipart==0.0.9